added basic get all and get individual routes for players and games
